feat: add LanguageOptions for language filtering and update imports in Languages
